Penultimate HW 01 update. Added lots of stuff.
Added JavaScript functionality that rearranges ordering of elements
in the header and content-boxes when screen size dips below a certain
pizel width. This is in order to facilitate for small windows sizes and/
or mobile devices. This is triggered once upon the page loading, and
subsequently upon any resize of the window. 

Lots of little changes to website layout. The next update will be the
final submission for HW 01, due on 1/20/2020